requirements.txt
streamlit
tensorflow
streamlit-webrtc
opencv-python-headless
pillow
numpy
scipy
